feat(#707) - fix attempting to delete not existing models

+++ application/app/Sync/Repositories/ClassifierValueRepository.php
@@ -29,8 +29,9 @@ class ClassifierValueRepository implements CachedEntityRepositoryInterface
 
     public function delete(string $id): void
     {
-        $obj = $this->getBaseModel()->find($id);
-        $obj->delete();
+        if ($obj = $this->getBaseModel()->find($id)) {
+            $obj->delete();
+        }
     }
 
     public function deleteNotSynced(): void
